requirements.txt
flask
scikit-learn
joblib
pandas
python-dotenv
google-generativeai
